fix: ensure styles can be overridden
The previous setup meant that if you were to provide your own `paddingBottom` to the style object on the `tabBarOptions`, it would not override the default value. For some people, the size of the tab bar initially may be too large to suit there UI so having the ability to override this is quite important.